pyproject.toml
[project]
name = "AMAO"
version = "0.1.0"
description = "Adaptive Multi-Agent Orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "chromadb>=1.5.4",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "motor>=3.7.1",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.3",
    "sqlmodel>=0.0.37",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
]
